IN-50 refactor: checkbox refactor
