+++ public/icons/index.ts
@@ -95,4 +95,4 @@ export {default as Paid} from './Paid';
 export {default as DoneAllOutline} from './DoneAllOutline';
 export {default as ArrowBackOutline} from './ArrowBackOutline'
 export {default as ArrowForwardOutline} from './ArrowForwardOutline'
-export {default as Checkmark} from './CheckMark'
+export {default as CheckmarkOutline} from './CheckmarkOutline'
