+++ shared/components/Сheckbox/Сheckbox.stories.tsx
@@ -1,11 +1,14 @@
 import { type Meta, type StoryObj } from '@storybook/react'
 import { useState } from 'react'
-import { Checkbox } from './Сheckbox'
+import { Checkbox } from '@/shared/components'
 
 const meta = {
   component: Checkbox,
   tags: ['autodocs'],
   title: 'Components/UI/Checkbox',
+  args: {
+    label: 'Checkbox',
+  },
   argTypes: {
     disabled: {
       control: 'boolean',
@@ -29,33 +32,34 @@ const meta = {
 export default meta
 type Story = StoryObj<typeof meta>
 
-export const Default: Story = {
+export const Checked: Story = {
   args: {
-    label: 'Check-Box',
     checked: true,
   },
 }
 
-export const RecaptchaMode: Story = {
+export const Unchecked: Story = {
   args: {
-    label: 'I’m not a robot',
     checked: false,
-    recaptchaMode: true,
   },
-  render: args => {
-    const [loading, setLoading] = useState(false)
-    const [verified, setVerified] = useState(false)
+}
 
-    const handleChange = async (checked: boolean) => {
-      setLoading(true)
-      await new Promise(resolve => setTimeout(resolve, 2000))
-      setLoading(false)
-      setVerified(true)
-      return true
-    }
-
-    return (
-      <Checkbox {...args} checked={verified} disabled={loading} onCheckedChange={handleChange} />
-    )
+export const Indeterminate: Story = {
+  args: {
+    checked: 'indeterminate',
   },
+}
+
+export const Controlled: Story = {
+  render: () => {
+    return <CheckboxWrapper />
+  },
+}
+
+function CheckboxWrapper() {
+  const [checked, setChecked] = useState(false)
+
+  return (
+    <Checkbox checked={checked} label="Controlled" onCheckedChange={() => setChecked(!checked)} />
+  )
 }
