+++ shared/components/Сheckbox/Сheckbox.tsx
@@ -1,61 +1,23 @@
 'use client'
 
-import { type ComponentPropsWithRef, useState, useId, type Ref } from 'react'
+import { type ComponentPropsWithRef, useId } from 'react'
 import * as CheckboxRadix from '@radix-ui/react-checkbox'
 import s from './Сheckbox.module.scss'
 import { clsx } from 'clsx'
-import CheckIcon from '../../../public/icons/CheckMark'
-import { Label } from '../Label/Label'
+import { CheckmarkOutline } from '@/public'
+import { Label } from '@/shared/components'
 
 export type CheckboxProps = {
   label?: string
-  onCheckedChange?: (checked: boolean) => Promise<boolean> | void
-  ref?: Ref<HTMLButtonElement>
-  recaptchaMode?: boolean
 } & ComponentPropsWithRef<typeof CheckboxRadix.Root>
 
-export const Checkbox = ({
-  disabled,
-  id,
-  label,
-  className,
-  checked: externalChecked,
-  onCheckedChange,
-  ref,
-  recaptchaMode = false,
-  ...rest
-}: CheckboxProps) => {
-  const [isLoading, setIsLoading] = useState(false)
-  const [isVerified, setIsVerified] = useState(false)
-  const [internalChecked, setInternalChecked] = useState(false)
+export const Checkbox = ({ disabled, id, label, className, checked, ...rest }: CheckboxProps) => {
   const useID = useId()
   const checkBoxID = id ?? useID
-
-  const checked = recaptchaMode ? internalChecked : externalChecked
-
-  const handleCheckedChange = async (newChecked: boolean) => {
-    if (recaptchaMode) {
-      setInternalChecked(true)
-    }
-
-    if (onCheckedChange) {
-      setIsLoading(true)
-      try {
-        const success = await onCheckedChange(newChecked)
-        setIsVerified(!!success)
-        if (recaptchaMode) {
-          setInternalChecked(!!success)
-        }
-      } finally {
-        setIsLoading(false)
-      }
-    }
-  }
-
   const classNames = {
     label: clsx(s.label, disabled && s.disabled),
     container: clsx(s.container, className),
-    root: clsx(s.root, recaptchaMode && s.recaptchaRoot, isLoading && s.hidden),
+    root: clsx(s.root),
     indicator: s.indicator,
     buttonWrapper: clsx(s.buttonWrapper, disabled && s.disabled),
     loader: s.loader,
@@ -63,24 +25,18 @@ export const Checkbox = ({
 
   return (
     <div className={classNames.container}>
-      <div className={classNames.buttonWrapper}>
-        {isLoading && recaptchaMode && <span className={classNames.loader}></span>}
-
-        <CheckboxRadix.Root
-          ref={ref}
-          checked={isVerified || checked}
-          className={classNames.root}
-          data-verified={isVerified}
-          disabled={disabled || isLoading}
-          id={checkBoxID}
-          onCheckedChange={handleCheckedChange}
-          {...rest}
-        >
-          <CheckboxRadix.Indicator className={classNames.indicator}>
-            {(isVerified || checked) && <CheckIcon />}
-          </CheckboxRadix.Indicator>
-        </CheckboxRadix.Root>
-      </div>
+      <CheckboxRadix.Root
+        {...rest}
+        checked={checked}
+        className={classNames.root}
+        disabled={disabled}
+        id={checkBoxID}
+      >
+        <CheckboxRadix.Indicator className={classNames.indicator}>
+          {checked === true && <CheckmarkOutline />}
+          {checked === 'indeterminate' && <span className={classNames.loader}></span>}
+        </CheckboxRadix.Indicator>
+      </CheckboxRadix.Root>
       {label && (
         <Label className={classNames.label} htmlFor={checkBoxID}>
           {label}
